requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
canvas==0.3
canvasapi==3.0.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-filter==22.1
django-json-field==0.5.7
django-jsonfield==1.4.1
django-jsonfield-backport==1.0.5
django-mysql==4.9.0
django-oauth-toolkit==2.2.0
django-paypal==2.0
django-picklefield==3.1
django-qr-code==3.1.1
django-rest-swagger==2.2.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==1.2.1
drf-yasg==1.21.5
ecdsa==0.18.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
filelock==3.9.0
Flask==2.2.2
idna==3.4
inflection==0.5.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jwcrypto==1.4.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
mysql-connector-python==8.0.30
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.4
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
paypalrestsdk==1.13.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
psycopg2==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.10.7
pygame==2.0.2
PyGetWindow==0.0.9
Pygments==2.12.0
PyJWT==2.6.0
PyMsgBox==1.0.9
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyperclip==1.8.2
pypng==0.20220715.0
PyQRCode==1.2.1
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytweening==1.0.4
pytz==2022.7.1
pywhatkit==5.4
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
qrcode==7.4.2
qtconsole==5.3.1
QtPy==2.1.0
redmail==0.6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.1
scipy==1.10.0
segno==1.5.2
Send2Trash==1.8.0
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.1.0
social-auth-core==4.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.3.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.19.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wikipedia==1.4.0
wrapt==1.15.0
